requirements.txt
# Horse Racing Data Analysis Project Dependencies

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine learning
tensorflow>=2.13.0
keras>=2.13.0
scikit-learn>=1.3.0

# Logging and configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Type hints
typing-extensions>=4.7.0 